refactor: move metaProxy onto LocalizationStringMetaData
This way we don't have to forward both and the meta proxy is only generated once per resource call,
which is good if a template script loads another resource with a template script that loads another,
etc. We don't need to make a proxy for each level of recursion.

With the public api of Localization.resource() no longer listing _meta in the parameters we won't
have to worry about it being passed. Regrettably this means casting Localization.resource to
when we need to pass _meta behind the scenes but it's a necessary sacrifice.