feat: allow one-line resources in localization files
* Clean up

While allowing one-line strings in localization files, this technically allows multi-line strings
where the first line begins on the same line as the localization key. Obviously this would be
ugly within the localization file so it should be avoided.

The only way to avoid this would be to peek the entire remainder of the node, including all
chunk kinds, and determine if the outcome only spans a single line. This would mean capturing
the entire raw string which means effectively making a separate pass over the entire string
prior to the actual parsing pass. This is undesirable from a performance standpoint for sure.

I might play around with it and run some benchmarks. Currently I want to avoid any overall
slowdown of the parser as much as possible.