refactor: generation of script template convenience vars
They're now generated based off of the script content itself so we're not allocating variables that aren't actually used (since some declared arguments might just be used in other templates and not script templates).

Because there is no static analysis of the code involved, anything that looks like a `$template_argument` will generate an associated variable as long as the identifier is valid. This means a string like `'$foo'` will generate the variable `$foo`, despite just being a string and not an actual usage of a template argument.